Moved test directories
