+++ jest.config.js
@@ -3,6 +3,6 @@ export default {
       '^.+\\.ts?$': 'ts-jest'
     },
     testEnvironment: 'node',
-    testRegex: '/(tests|runtime-tests)/.*\\.test?\\.(ts|js)$',
+    testRegex: '/tests/(plugin|runtime)/.*\\.test?\\.(ts|js)$',
     moduleFileExtensions: ['ts', 'js']
 };
