+++ tests/plugin/helpers-commonjs/descriptionHelper.d.ts
