+++ tests/plugin/helpers-commonjs/descriptionHelper.js
