+++ tests/plugin/helpers/image.js
