+++ tests/plugin/helpers2/json.js
