+++ tests/plugin/index.test.ts
@@ -1,67 +1,11 @@
 import fs from 'fs';
 import path from 'path';
-import { js_beautify } from 'js-beautify';
 
-import HandlebarsTransformer from '../lib/handlebars-transformer';
-
-import { HandlebarsPluginOptions } from '../lib/types/plugin-options';
-import { CompileResult } from '../lib/types/handlebars';
+import { HandlebarsPluginOptions } from '../../lib/types/plugin-options';
 import { lookupHelperRegistration, lookupPartialRegistration } from './utils/utils';
+import { removeOutputDir, testTemplate } from './utils';
 
-type TestFn = (err: Error, output: CompileResult) => {}
-
-const testFunctionsDir = '../runtime-tests/functions'
-const absoluteTestFunctionsDir = path.resolve(__dirname, testFunctionsDir)
-if (fs.existsSync(absoluteTestFunctionsDir)) {
-    fs.rmdirSync(absoluteTestFunctionsDir, { recursive: true });
-}
-
-function loadTemplate(templatePath: string) {
-    return fs.readFileSync(templatePath).toString();
-}
-
-function createFunctionFile(template: string, output: CompileResult) {
-    const templatePath = path.join(__dirname, template)
-    const extname = path.extname(template)
-    const outputTemplate = path.join(__dirname, testFunctionsDir, template)
-    const outputDir = path.dirname(outputTemplate)
-    const outputName = `${path.basename(templatePath, extname)}.js`
-    const outputPath = path.join(outputDir, outputName)
-    if (!fs.existsSync(outputDir)) {
-        fs.mkdirSync(outputDir, { recursive: true });
-    }
-    let code: string = output?.code || ''
-    code = !code ? code : js_beautify(code)
-    fs.writeFileSync(outputPath, code)
-}
-
-function testTemplate(template: string, pluginOptions: HandlebarsPluginOptions, testFn: TestFn) {
-
-    let err: Error | undefined
-    const hbsTransformer = new HandlebarsTransformer(pluginOptions)
-
-    const templatePath = path.join(__dirname, template)
-    
-    let source: string | null = null
-    try {
-        source = loadTemplate(templatePath)
-    } catch (e) {
-        err = e
-        return
-    }
-
-    let output: CompileResult | null = null
-    try {
-        output = hbsTransformer.transform(source, templatePath)
-    } catch (e) {
-        err = e
-    }
-
-    testFn(err, output)
-
-    createFunctionFile(template, output)
-
-}
+removeOutputDir()
 
 describe('Handlebars Transformer', () => {
