+++ tests/plugin/invalid-syntax-error.hbs
