+++ tests/plugin/nested-templates/nested/with-ancestor-dir-partial.hbs
