+++ tests/plugin/nested-templates/with-cousin-dir-partial.hbs
