+++ tests/plugin/nested-templates/with-parent-dir-partial.hbs
