+++ tests/plugin/partialDirs/anotherDir/otherPartial.hbs
