+++ tests/plugin/partialDirs/otherPartial.hbs
