+++ tests/plugin/some-partial.hbs
