+++ tests/plugin/with-dir-helpers.hbs
