+++ tests/plugin/with-dir-partials.hbs
