+++ tests/plugin/with-excluded-helpers.hbs
