+++ tests/plugin/with-helpers-commonjs.hbs
