+++ tests/plugin/with-inline-partial.hbs
