+++ tests/plugin/with-inline-requires.hbs
