+++ tests/plugin/with-known-helpers.hbs
