+++ tests/plugin/with-partial-block.hbs
