+++ tests/plugin/with-partials-helpers-inline-requires.hbs
