+++ tests/plugin/with-partials.hbs
