+++ tests/plugin/with-plugin-partial.hbs
