+++ tests/plugin/with-relative-root.hbs
